update schema generation script
